requirements.txt
pandas
numpy
matplotlib
scikit-learn
scipy
sqlalchemy
snowflake-sqlalchemy
streamlit
joblib
gdown
